Added support for tcp_taps new TCP_TAP_NICNAME
tcp_tap's TAP_LOG_NAME adds much better ability to control incoming
socket-binds. That was part of a bug-fix, but a very nice side-effect is
that security can be moderated, from accepting only 127.0.0.1 (impossible
for outsiders to spoof) into "@ANY@", "@HOSTNAME@" or specific name, each
which their particular use-cases. Small fix did all that (this message is
actually much longer)

Additionally:
* Changes in profile-names
* Log-format change. Contains also PID now, should more then one sessions
  run.

Note: there is now a new hight-level log. It's low traffic and can be left
on. It also defaults to /tmp/something (i.e. will be cleared upon restart).
File-name is controlled via profile-files, but nothing is printed until
profiles are read which is why it works. Filename for this log is controlled
by env-var TAP_LOG_NAME.

Change-Id: I70705f2a73b46a22caea65b7f32bc270735e4de0