Added first lines of nc and rlwrapper code
What beautiful pieces of code they are! Remote session now works almost as
well as if it would be a native GDB session. That with practically *no
change to tcp_tap at all*! Somewhat different than with a real gdb-session
but now much. This really made an enormous difference and it proves me right
not to add that functionality to tcp_tap. It would only have have made it
more complicated and bloated with non-generic code. Also as a side-effect:
It would had made it more GDB  specific (which I don't want).

Having both screen, rlwrapper and nc as helpers give:

* Infinite scroll-back [screen]
* Copy&paste from either your's, the GUI's or other users (sessions) into
  your own execution [screen]
* Memorizing other session users operations into your own history [screen] +
  [rlwrapper].
* Memory of history even across sessions (great!) [rlwapper/readline]
* Potential TAB fill-out. Need to research this more [rlwapper/readline]
* Acting on signals as they were intended to [nc]. No special secret command
  needed any more to exit group sessions. Just ^C and you're done.

nc and rlwrapper now completely replace telnet and is from now-on the
preferred method of connection. This commit contains a fullback to telnet if
both of these aren't installed on your system however. Will write a small
client using these for remote users as well.

Transparency is King(k)!

Additionally:
* Corrected bug about handling local IF's (that's what the diffs about, the
  above is just me rambling about at a late hour with a glass of wine to
  celebrate :))

Change-Id: Ida8f46346513769fc5e841119bf9db95008ef192